pyproject.toml
[project]
name = "rf-spectrum-env"
version = "1.0.0"
description = "RF Spectrum Allocation OpenEnv Environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "pydantic>=2.0",
    "uvicorn>=0.27.0",
    "fastapi>=0.109.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rf_spectrum_env*", "models*", "scenarios*", "client*", "server*"]
